pyproject.toml
[project]
name = "trilium-sender-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=22.5",
    "ruff>=0.14.9",
    "trilium-py>=1.3.2",
]
